+++ source/Button/index.js
@@ -4,3 +4,4 @@
  */
 
 export Button from './Button';
+export ButtonGroup from './ButtonGroup';