+++ source/index.js
@@ -8,4 +8,5 @@ export HorizontalFlex from './HorizontalFlex';
 export VerticalFlex from './VerticalFlex';
 export Toast from './Toast';
 export List from './List';
+export Button from './Button';
 export colors from './colors';